+++ src/main/java/org/hyejoon/cuvcourse/domain/course/courseregist/service/lock/DistributedLock.java
@@ -1,4 +1,4 @@
-package org.hyejoon.cuvcourse.domain.course.courseregist.service;
+package org.hyejoon.cuvcourse.domain.course.courseregist.service.lock;
 
 public interface DistributedLock {
